requirements.txt
discord.py
requests
psycopg2-binary
